requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate
Flask-WTF
WTForms
python-slugify==8.0.4
